% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detect_missing_value.R
\name{detect_missing_values}
\alias{detect_missing_values}
\title{Detect Missing Values}
\usage{
detect_missing_values(data)
}
\arguments{
\item{data}{A data frame to analyze for missing values.}
}
\value{
A named numeric vector where each name is a column with missing values, and the value is the count of missing values in that column.
}
\description{
This function calculates the number of missing values (NA) in each column of a dataset.
}
\examples{
data(airquality)
detect_missing_values(airquality)
}
