% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_Missing_Value.R
\name{replace_missing_with_median}
\alias{replace_missing_with_median}
\title{Replace Missing Values with Medians}
\usage{
replace_missing_with_median(data)
}
\arguments{
\item{data}{A data frame with numeric columns containing missing values.}
}
\value{
A data frame with missing values replaced by column medians.
}
\description{
This function replaces all missing values in numeric columns with the median of their respective columns.
}
\examples{
data(airquality)
cleaned_airquality <- replace_missing_with_median(airquality)
print(cleaned_airquality)
}
